Move main context filtering logic to sentry-rails
